requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
apt-clone==0.2.1
apturl==0.5.2
async-timeout==4.0.3
attrs==23.1.0
av==10.0.0
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.28.62
botocore==1.31.62
Brlapi==0.8.3
ccsm==0.9.14.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
compizconfig-python==0.9.14.1
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
customtkinter==5.2.0
darkdetect==0.8.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
eyeD3==0.8.10
filelock==3.6.0
fpdf==1.7.2
frozenlist==1.4.0
grpcio==1.30.2
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
imageio==2.31.5
imageio-ffmpeg==0.4.9
IMDbPY==2021.4.18
importlib-metadata==4.6.4
jeepney==0.7.1
jmespath==1.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
moviepy==1.0.3
multidict==6.0.4
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.0
oauthlib==3.2.0
onboard==1.4.1
openai==0.28.1
packaging==21.3
pexpect==4.8.0
Pillow==9.0.1
proglog==0.1.10
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyelftools==0.27
pyglet==2.0.9
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
s3transfer==0.7.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
systemd-python==234
tkvideoplayer==2.3
tldextract==3.1.2
tqdm==4.66.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
xdg==5
xkit==0.0.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==1.0.0
